fix: database.yml 수정
