+++ backend/src/main/java/com/carrot/global/error/ErrorCode.java
@@ -11,6 +11,7 @@ public enum ErrorCode {
     EMAIL_VALIDATION_ERROR(BAD_REQUEST, "이메일 형식이 잘못되었습니다."),
     NICKNAME_LENGTH_VALIDATION_ERROR(BAD_REQUEST, "닉네임의 길이가 15을 초과했습니다."),
     USER_NOTFOUND_ERROR(NOT_FOUND, "존재하지 않는 회원입니다."),
+    USER_REGION_NOT_SAVE_REGION(UNAUTHORIZED, "저장된 지역 정보가 존재하지 않습니다"),
     USER_REGION_MAX_ERROR(BAD_REQUEST, "지역정보는 최대 2개까지 저장할 수 있습니다."),
     USER_REGION_NOTFOUND_ERROR(NOT_FOUND, "존재하지 않는 회원의 지역정보입니다."),
     USER_REGION_VALIDATION_ERROR(UNAUTHORIZED, "해당 지역정보에 관한 권한이 없습니다."),
