+++ backend/src/test/resources/application-database.yml
@@ -24,5 +24,5 @@ spring:
 
     session:
       host: localhost
-      port: 6380
+      port: 6379
       password: password
